feat: return timestamp from `GetIDs` (#103)
* feat: return timestamp from `GetIDs`

Enhanced the GetIDs functionality to include a timestamp in the response, both in the proto files and Go structures. Updated relevant files and added checks for the new field in the tests.

* chore: fix linter errors

* fix: remove unused struct

* chore: Switch from Docker to 'plain' Buf for protobuf generation

Removed old protobuf generation shell scripts and updated Makefile to use Buf. Changed dependencies and updated configurations accordingly. This modernizes the setup and removes the dependency on Docker.

* refactor: use google.protobuf.Timestamp instead of int64

Replaced the int64 timestamp with google.protobuf.Timestamp in the GetIdsResponse message. Adjusted the generated code to handle the new timestamp type and updated the protobuf imports accordingly.

* ci: fix PATH issues

* ci: use Go version from go.mod